+++ CHANGELOG.md
@@ -25,10 +25,11 @@ To Document:
 * Change in behavior: beforeAll() and afterAll() now notify upon completion (change name of onTestCaseComplete? To onProgress maybe?)
 * Change in behavior: beforeAll() and afterAll() failures no longer result in failing TestCaseResults. Instead, they show up in TestSuiteResult.beforeAll or .afterAll
 * Change in behavior: beforeAll() and afterAll() are picked up by allTests(), allMatchingTests(), and allMatchingMarks() (as 'not marked')
+* Change in behavior: Runs all afterAll() blocks even if one fails 
 
 TO DO:
-* Fix beforeAll/afterAll failure reporting; currently notifies, but doesn't show up in test summary
-* Decide how to report on before/after all/each failures (all currently gets reported as a test, but only when it fails; each overrides actual test result)
+* beforeAll() / afterAll() shouldn't put number in name until second one (like beforeEach())
+* Decide how to report on beforeEach/afterEach failures (overrides actual test result)
 
 
 ## v0.11.x: Add optional 'actual' and 'expected' to assert.fail()